// src/components/Types.ts
export type QuesionsType = {

    "id": number,

    "quesion": string;

    "answer-1": string;

    "answer-2": string;

    "answer-3": string;

    "answer-4": string;

    "right-anwser": string;

}

export type StateType = {

    "answer-1": string;

    "answer-2": string;

    "answer-3": string;

    "answer-4": string;

    isClicked: boolean;

    score: number;

}